PeUHqcqv: split out the responder status code translation functions
