+++ src/main/java/uk/gov/ida/verifyserviceprovider/services/AssertionService.java
@@ -1,7 +1,6 @@
 package uk.gov.ida.verifyserviceprovider.services;
 
 import org.opensaml.saml.saml2.core.Assertion;
-import org.opensaml.saml.saml2.core.StatusCode;
 import uk.gov.ida.verifyserviceprovider.dto.LevelOfAssurance;
 import uk.gov.ida.verifyserviceprovider.dto.TranslatedResponseBody;
 
@@ -15,7 +14,4 @@ public interface AssertionService {
             LevelOfAssurance expectedLevelOfAssurance,
             String entityId
     );
-
-    TranslatedResponseBody translateNonSuccessResponse( StatusCode statusCode);
-
 }
