+++ src/main/java/uk/gov/ida/verifyserviceprovider/services/ResponderResponseTranslator.java
@@ -1,0 +1,8 @@
+package uk.gov.ida.verifyserviceprovider.services;
+
+import org.opensaml.saml.saml2.core.StatusCode;
+import uk.gov.ida.verifyserviceprovider.dto.TranslatedResponseBody;
+
+public interface ResponderResponseTranslator {
+    TranslatedResponseBody translateResponderResponse(StatusCode statusCode);
+}
