+++ src/test/java/unit/uk/gov/ida/verifyserviceprovider/services/MsaAssertionServiceTest.java
@@ -10,7 +10,6 @@ import org.junit.rules.ExpectedException;
 import org.opensaml.saml.saml2.core.Assertion;
 import org.opensaml.saml.saml2.core.AuthnStatement;
 import org.opensaml.saml.saml2.core.Issuer;
-import org.opensaml.saml.saml2.core.StatusCode;
 import org.opensaml.security.credential.BasicCredential;
 import org.opensaml.security.credential.Credential;
 import org.opensaml.security.credential.impl.CollectionCredentialResolver;
@@ -18,7 +17,6 @@ import org.opensaml.security.crypto.KeySupport;
 import org.opensaml.xmlsec.config.impl.DefaultSecurityConfigurationBootstrap;
 import org.opensaml.xmlsec.signature.support.impl.ExplicitKeySignatureTrustEngine;
 import uk.gov.ida.saml.core.IdaSamlBootstrap;
-import uk.gov.ida.saml.core.domain.SamlStatusCode;
 import uk.gov.ida.saml.core.test.PrivateKeyStoreFactory;
 import uk.gov.ida.saml.core.test.TestCredentialFactory;
 import uk.gov.ida.saml.core.test.TestEntityIds;
@@ -27,7 +25,6 @@ import uk.gov.ida.saml.core.test.builders.ConditionsBuilder;
 import uk.gov.ida.saml.core.test.builders.IssuerBuilder;
 import uk.gov.ida.saml.core.test.builders.SubjectBuilder;
 import uk.gov.ida.saml.core.validation.SamlTransformationErrorException;
-import uk.gov.ida.verifyserviceprovider.dto.MatchingScenario;
 import uk.gov.ida.verifyserviceprovider.dto.TranslatedMatchingResponseBody;
 import uk.gov.ida.verifyserviceprovider.exceptions.SamlResponseValidationException;
 import uk.gov.ida.verifyserviceprovider.factories.saml.ResponseFactory;
@@ -44,10 +41,7 @@ import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.ida.saml.core.extensions.IdaAuthnContext.LEVEL_2_AUTHN_CTX;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_PRIVATE_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_PUBLIC_CERT;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_MS_PRIVATE_SIGNING_KEY;
-import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_MS_PUBLIC_SIGNING_CERT;
+import static uk.gov.ida.saml.core.test.TestCertificateStrings.*;
 import static uk.gov.ida.saml.core.test.builders.AssertionBuilder.anAssertion;
 import static uk.gov.ida.saml.core.test.builders.AudienceRestrictionBuilder.anAudienceRestriction;
 import static uk.gov.ida.saml.core.test.builders.AuthnContextBuilder.anAuthnContext;
@@ -55,7 +49,6 @@ import static uk.gov.ida.saml.core.test.builders.AuthnContextClassRefBuilder.anA
 import static uk.gov.ida.saml.core.test.builders.AuthnStatementBuilder.anAuthnStatement;
 import static uk.gov.ida.saml.core.test.builders.ConditionsBuilder.aConditions;
 import static uk.gov.ida.saml.core.test.builders.SignatureBuilder.aSignature;
-import static uk.gov.ida.saml.core.test.builders.StatusCodeBuilder.aStatusCode;
 import static uk.gov.ida.saml.core.test.builders.SubjectBuilder.aSubject;
 import static uk.gov.ida.saml.core.test.builders.SubjectConfirmationBuilder.aSubjectConfirmation;
 import static uk.gov.ida.saml.core.test.builders.SubjectConfirmationDataBuilder.aSubjectConfirmationData;
@@ -216,67 +209,6 @@ public class MsaAssertionServiceTest {
             .buildUnencrypted();
 
         msaAssertionService.translateSuccessResponse(ImmutableList.of(assertion), IN_RESPONSE_TO, LEVEL_2, VERIFY_SERVICE_PROVIDER_ENTITY_ID);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenNonSuccessResponseCalledWithNoSubStatusCode() {
-        expectedException.expect(SamlResponseValidationException.class);
-        expectedException.expectMessage("Missing status code for non-Success response");
-
-        StatusCode statusCode = aStatusCode().withValue(StatusCode.RESPONDER).build();
-        msaAssertionService.translateNonSuccessResponse(statusCode);
-    }
-
-    @Test
-    public void shouldReturnScenarioCancelledWhenNoAuthnContextStatus() {
-        StatusCode statusCode = aStatusCode()
-                .withValue(StatusCode.RESPONDER)
-                .withSubStatusCode(aStatusCode().withValue(StatusCode.NO_AUTHN_CONTEXT).build())
-                .build();
-        TranslatedMatchingResponseBody response = msaAssertionService.translateNonSuccessResponse(statusCode);
-        assertThat(response.getScenario()).isEqualTo(MatchingScenario.CANCELLATION);
-    }
-
-    @Test
-    public void shouldReturnScenarioNoMatchWhenNoMatchStatus() {
-        StatusCode statusCode = aStatusCode()
-                .withValue(StatusCode.RESPONDER)
-                .withSubStatusCode(aStatusCode().withValue(SamlStatusCode.NO_MATCH).build())
-                .build();
-        TranslatedMatchingResponseBody response = msaAssertionService.translateNonSuccessResponse(statusCode);
-        assertThat(response.getScenario()).isEqualTo(MatchingScenario.NO_MATCH);
-    }
-
-    @Test
-    public void shouldReturnScenarioAuthenticationFailedWhenAuthnFailedStatus() {
-        StatusCode statusCode = aStatusCode()
-                .withValue(StatusCode.RESPONDER)
-                .withSubStatusCode(aStatusCode().withValue(StatusCode.AUTHN_FAILED).build())
-                .build();
-        TranslatedMatchingResponseBody response = msaAssertionService.translateNonSuccessResponse(statusCode);
-        assertThat(response.getScenario()).isEqualTo(MatchingScenario.AUTHENTICATION_FAILED);
-    }
-
-    @Test
-    public void shouldReturnScenarioRequestErrorWhenRequesterStatus() {
-        StatusCode statusCode = aStatusCode()
-                .withValue(StatusCode.RESPONDER)
-                .withSubStatusCode(aStatusCode().withValue(StatusCode.REQUESTER).build())
-                .build();
-        TranslatedMatchingResponseBody response = msaAssertionService.translateNonSuccessResponse(statusCode);
-        assertThat(response.getScenario()).isEqualTo(MatchingScenario.REQUEST_ERROR);
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenNonSuccessResponseCalledWithUnrecognisedStatus() {
-        expectedException.expect(SamlResponseValidationException.class);
-        expectedException.expectMessage("Unknown SAML sub-status: urn:oasis:names:tc:SAML:2.0:status:NoAvailableIDP");
-
-        StatusCode statusCode = aStatusCode()
-                .withValue(StatusCode.RESPONDER)
-                .withSubStatusCode(aStatusCode().withValue(StatusCode.NO_AVAILABLE_IDP).build())
-                .build();
-        msaAssertionService.translateNonSuccessResponse(statusCode);
     }
 
     private AssertionBuilder aSignedAssertion() {
